# specs/openapi.yaml
openapi: 3.1.0
info:
  title: ScrumQuest API
  version: 1.0.0
  description: |
    Real-time multiplayer scrum poker estimation game API.

    ScrumQuest provides REST endpoints for authentication, user profiles,
    and health checks. Real-time game functionality is handled via WebSocket
    (Socket.IO) connections documented separately.
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://scrumquest.app
    description: Production server (placeholder)
...
paths:
...
  /api/auth/register:
    post:
      operationId: registerUser
      summary: Register new user
      description: Create a new user account with email and password
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: john@example.com
              password: securepassword123
              username: johndoe
              displayName: John Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingFields:
                  summary: Missing required fields
                  value:
                    error: Email and password are required
                shortPassword:
                  summary: Password too short
                  value:
                    error: Password must be at least 8 characters
                emailTaken:
                  summary: Email already registered
                  value:
                    error: Email already registered
                usernameTaken:
                  summary: Username taken
                  value:
                    error: Username already taken
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/login:
    post:
      operationId: loginUser
      summary: Login with email/password
      description: Authenticate user with email and password credentials
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: john@example.com
              password: securepassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid credentials
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/logout:
    post:
      operationId: logoutUser
      summary: Logout current user
      description: End the current session and clear authentication cookies
      tags:
        - auth
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    const: true
              example:
                success: true
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after login
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - email
        - displayName
      properties:
        id:
          type: integer
          description: Unique user identifier
        username:
          type: string
          description: Unique username
        email:
          type: string
          format: email
          description: User's email address
        displayName:
          type: string
          description: Display name shown in game
        avatarUrl:
          type: string
          format: uri
          nullable: true
          description: URL to user's avatar image
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: Password (minimum 8 characters)
        username:
          type: string
          description: Optional username (defaults to email prefix)
        displayName:
          type: string
          description: Optional display name (defaults to username)
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Operation failed
...